🚚 Support for trailers and trailer modification (#327)
This PR adds support for trailers.

Trailers can be extremely useful for gRPC connections, and this PR allows trailers from a backend connection to flow back through the system. This usage is validated in tests, and has been independently validated by examples on my laptop. In addition, it provides experimental support for viewing, modifying, and/or adding trailers through a new set of ABI functions.

---------

Co-authored-by: Mike Zeller <mzeller@fastly.com>
Co-authored-by: Adam C. Foltzer <acfoltzer@acfoltzer.net>