Add device detection hostcall (#330)
* Add hostcall for device_detection.

This is closely patterned on the geolocation hostcall.

The new test that executes the hostcall needs a new SDK, so this
will be followed by a separate commit that comments it out. That
can be reverted once the new SDK is available.

* device_detection: Temporarily comment out tests that need new SDK.

These tests depend on a new SDK for the device_detection hostcall,
comment them out for now. This is in its own commit to make it
easy to revert.

* Minor changes from cargo fmt.

* Fix copy/pasted identifier.